+++ src/Connection.php
@@ -15,7 +15,7 @@ class Connection
     protected $facebook_request;
 
     /**
-     * Response value object
+     * Response value object.
      *
      * @var \SammyK\FacebookQueryBuilder\Response
      */
@@ -29,18 +29,25 @@ class Connection
     protected static $facebook_session;
 
     /**
+     * The Facebook SDK session object using an app access token.
+     *
+     * @var \Facebook\FacebookSession
+     */
+    protected static $facebook_app_session;
+
+    /**
      * Facebook app ID.
      *
      * @var int
      */
-    protected static $app_id;
+    public static $app_id;
 
     /**
      * Facebook app secret.
      *
      * @var string
      */
-    protected static $app_secret;
+    public static $app_secret;
 
     /**
      * Access token to use for API calls.
@@ -78,16 +85,16 @@ class Connection
     /**
      * Sets the access token to be used for all API requests.
      *
-     * @param string $access_token
+     * @param AccessToken|string $access_token
      */
     public static function setAccessToken($access_token)
     {
-        static::$access_token = $access_token;
+        static::$access_token = (string) $access_token;
         static::$facebook_session = null; // Reset the FacebookSession
     }
 
     /**
-     * Get the existing FacebookSession object or new up one
+     * Get the existing FacebookSession object or new up one.
      *
      * @return \Facebook\FacebookSession
      */
@@ -99,7 +106,7 @@ class Connection
     }
 
     /**
-     * Set the FacebookSession object
+     * Set the FacebookSession object.
      *
      * @param \Facebook\FacebookSession $facebook_session
      */
@@ -109,9 +116,32 @@ class Connection
     }
 
     /**
+     * Get the existing FacebookSession object using an app access token or new up one.
+     *
+     * @return \Facebook\FacebookSession
+     */
+    public static function getFacebookAppSession()
+    {
+        if (isset(static::$facebook_app_session)) return static::$facebook_app_session;
+
+        return static::$facebook_app_session = new FacebookSession(static::$app_id . '|' . static::$app_secret);
+    }
+
+    /**
+     * Set the FacebookSession object using an app access token.
+     *
+     * @param \Facebook\FacebookSession $facebook_app_session
+     */
+    public static function setFacebookAppSession(FacebookSession $facebook_app_session)
+    {
+        static::$facebook_app_session = $facebook_app_session;
+    }
+
+    /**
      * Send GET request to Facebook API.
      *
      * @param \SammyK\FacebookQueryBuilder\RootEdge $root_edge
+     *
      * @return \SammyK\FacebookQueryBuilder\Response
      */
     public function get(RootEdge $root_edge)
@@ -124,6 +154,7 @@ class Connection
      *
      * @param \SammyK\FacebookQueryBuilder\RootEdge $root_edge
      * @param array $data
+     *
      * @return \SammyK\FacebookQueryBuilder\Response
      */
     public function post(RootEdge $root_edge, $data)
@@ -135,6 +166,7 @@ class Connection
      * Send DELETE request to Facebook API.
      *
      * @param \SammyK\FacebookQueryBuilder\RootEdge $root_edge
+     *
      * @return \SammyK\FacebookQueryBuilder\Response
      */
     public function delete(RootEdge $root_edge)
@@ -145,17 +177,22 @@ class Connection
     /**
      * Send a request to the Facebook Graph API.
      *
-     * @throws \SammyK\FacebookQueryBuilder\FacebookQueryBuilderException
      * @param string $path
      * @param string $method
      * @param array $params
+     * @param boolean $app_request
+     *
      * @return \Facebook\FacebookResponse
+     *
+     * @throws \SammyK\FacebookQueryBuilder\FacebookQueryBuilderException
      */
-    public function send($path, $method = 'GET', array $params = [])
+    public function send($path, $method = 'GET', array $params = [], $app_request = false)
     {
         try
         {
-            $request = $this->facebook_request->make(static::getFacebookSession(), $method, $path, $params);
+            $session = $app_request ? static::getFacebookAppSession() : static::getFacebookSession();
+
+            $request = $this->facebook_request->make($session, $method, $path, $params);
             $response = $request->execute();
 
             return $this->response->create($response);