+++ src/Response.php
@@ -29,10 +29,21 @@ class Response extends Collection
 
         if ($raw_response instanceof FacebookResponse)
         {
-            $raw_response = json_decode($raw_response->getRawResponse(), true);
+            // If JSON returned
+            $decoded_response = json_decode($raw_response->getRawResponse(), true);
+
+            // If kay/value pairs returned
+            if ($decoded_response === null)
+            {
+                parse_str($raw_response->getRawResponse(), $decoded_response);
+            }
+        }
+        else
+        {
+            $decoded_response = $raw_response;
         }
 
-        $this->raw_response = ! is_array($raw_response) ? [$raw_response] : $raw_response;
+        $this->raw_response = ! is_array($decoded_response) ? [$decoded_response] : $decoded_response;
 
         $this->response = static::castGraphObjects($this->raw_response);
     }
@@ -72,6 +83,7 @@ class Response extends Collection
      * Iterate response data recursively and cast to Graph value objects.
      *
      * @param array $data
+     *
      * @return \SammyK\FacebookQueryBuilder\Collection
      */
     public static function castGraphObjects(array $data)
@@ -82,10 +94,26 @@ class Response extends Collection
         }
         elseif (isset($data['data']))
         {
-            //$collection = $this->castGraphObjects($data['data']);
-            return new GraphCollection($data);
+            if (static::isCastableAsCollection($data['data']))
+            {
+                return new GraphCollection($data);
+            }
+            return new GraphObject($data['data']);
         }
 
         return new GraphObject($data);
+    }
+
+    /**
+     * Determines whether or not the data should be cast as a GraphCollection.
+     *
+     * @param array $data
+     *
+     * @return boolean
+     */
+    public static function isCastableAsCollection(array $data)
+    {
+        // Checks for a sequential numeric array which would be a GraphCollection
+        return array_keys($data) === range(0, count($data) - 1);
     }
 }