+++ tests/FacebookQueryBuilderExceptionTest.php
@@ -3,29 +3,44 @@
 use Mockery as m;
 use SammyK\FacebookQueryBuilder\FacebookQueryBuilderException;
 
+class FbqFooException extends \Exception {}
+
 class FacebookQueryBuilderExceptionTest extends PHPUnit_Framework_TestCase
 {
-    public function testCanInstantiateWithMessageAndCode()
+    public function tearDown()
     {
-        $e = new FacebookQueryBuilderException('Foo message', 123);
+        m::close();
+    }
+
+    /** @test */
+    public function an_exception_can_be_thrown_with_a_custom_message_and_code_and_previous_exception()
+    {
+        $previous_exception = m::mock('FbqFooException');
+
+        $e = new FacebookQueryBuilderException('Foo message', 123, $previous_exception);
 
         $this->assertEquals('Foo message', $e->getMessage());
         $this->assertEquals(123, $e->getCode());
+        $this->assertEquals($previous_exception, $e->getPrevious());
     }
 
-    public function testCanInstantiateWithFacebookSdkException()
+    /** @test */
+    public function an_exception_can_be_thrown_with_a_previous_facebook_sdk_exception()
     {
         $exception_mock = $this->setUpApiException(123, 'OAuthException', 'Foo message');
 
         $e = new FacebookQueryBuilderException($exception_mock);
 
         $this->assertInstanceOf('\SammyK\FacebookQueryBuilder\GraphError', $e->getResponse());
-        $this->assertEquals('Error communicating with Facebook: Foo message', $e->getMessage());
-        $this->assertEquals(123, $e->getCode());
+        $this->assertEquals('Graph returned an error response.', $e->getMessage());
+        $this->assertEquals('Foo message', $e->getPrevious()->getMessage());
+        $this->assertEquals(10, $e->getCode());
+        $this->assertEquals(123, $e->getPrevious()->getCode());
         $this->assertEquals('OAuthException', $e->getType());
     }
 
-    public function testErrorCodeHandling()
+    /** @test */
+    public function error_codes_can_be_translated_into_error_summaries()
     {
         $exception_mock = $this->setUpApiException(102);
         $this->runTheApiExceptionTest($exception_mock, 'Login required.');
@@ -52,7 +67,8 @@ class FacebookQueryBuilderExceptionTest extends PHPUnit_Framework_TestCase
         $this->runTheApiExceptionTest($exception_mock, 'Extended permission required.');
     }
 
-    public function testExtendedPermissionDetecting()
+    /** @test */
+    public function permissions_from_error_messages_can_be_extracted_into_an_array()
     {
         $exception_mock = $this->setUpApiException(298, 'OAuthException', '(#298) Requires extended permission: read_mailbox');
 
@@ -93,4 +109,5 @@ class FacebookQueryBuilderExceptionTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals($expected_summary, $e->errorSummary());
     }
+
 }